ResponseEntity on user

+++ back/src/main/java/com/jeniti/back/controller/UserController.java
@@ -5,9 +5,11 @@ import com.jeniti.back.entity.User_class;
 import com.jeniti.back.service.ChannelService;
 import com.jeniti.back.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.http.HttpResponse;
 import java.util.Optional;
 
 @RestController
@@ -28,17 +30,17 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User_class getUserById(@PathVariable("id") final long id) {
+    public ResponseEntity<User_class> getUserById(@PathVariable("id") final long id) {
         Optional<User_class> u =  uService.getUser(id);
         if (u.isPresent()) {
-            return u.get();
+            return ResponseEntity.ok(u.get());
         } else {
-            return null;
+            return ResponseEntity.notFound().build();
         }
     }
 
     @PutMapping("/{id}")
-    public User_class updateUser(@PathVariable("id") final long id, @RequestBody User_class user) {
+    public ResponseEntity<User_class> updateUser(@PathVariable("id") final long id, @RequestBody User_class user) {
         Optional<User_class> u = uService.getUser(id);
         if (u.isPresent()) {
             User_class currentUser = u.get();
@@ -60,22 +62,31 @@ public class UserController {
             }
 
             Channel current_channel = user.getCurrent_channel();
-            if (current_channel != null && current_channel.getId() != null) {
-                Long current_channel_id = current_channel.getId();
-                Optional<Channel> c = cService.getByIdChannel(current_channel_id);
-                if (c.isPresent()) {
-                    currentUser.setCurrent_channel(c.get());
+            if (current_channel != null) {
+                if (current_channel.getId() != null) {
+                    Long current_channel_id = current_channel.getId();
+                    Optional<Channel> c = cService.getByIdChannel(current_channel_id);
+                    if (c.isPresent()) {
+                        currentUser.setCurrent_channel(c.get());
+                    }
+                } else {
+                    ResponseEntity.badRequest().build();
                 }
             }
-
-            return uService.UpdateUser(currentUser);
+            return ResponseEntity.ok(uService.UpdateUser(currentUser));
         } else {
-            return null;
+            return ResponseEntity.notFound().build();
         }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable("id") final long id) {
-        uService.deleteUser(id);
+    public ResponseEntity deleteUser(@PathVariable("id") final long id) {
+        Optional<User_class> u = uService.getUser(id);
+        if (u.isPresent()) {
+            uService.deleteUser(id);
+            return ResponseEntity.ok("L'utilisateur n°" + id + " a bien été supprimé");
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
